fix invalid check of rc for desp mlist
