+++ sample/filemap.c
@@ -12,6 +12,7 @@
 #include "asmdiocommon.h"
 #include "util.h"
 #include "dio.h"
+#include "mem.h"
 #include "iosvcs.h"
 #include "fm.h"
 #include "fmopts.h"
