+++ src/Makefile
@@ -1,4 +1,4 @@
-all: call31.o call31a.o dio.o s99.o dioa.o ihadcb.o util.o iosvcs.o
+all: call31.o call31a.o dio.o mem.o s99.o dioa.o mema.o ihadcb.o util.o iosvcs.o
 
 clean:
 	rm -f *.o *.lst *.dbg
@@ -19,6 +19,9 @@ call31a.o: call31a.s
 dio.o: dio.c $(IINCDIR)/wrappers.h $(INCDIR)/*.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -I$(IINCDIR) -c $<
 
+mem.o: mem.c $(IINCDIR)/wrappers.h $(INCDIR)/*.h
+	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -I$(IINCDIR) -c $<
+
 s99.o: s99.c $(IINCDIR)/wrappers.h $(INCDIR)/dio.h $(INCDIR)/s99.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -I$(IINCDIR) -c $<
 
@@ -35,4 +38,7 @@ ihadcb.o: ihadcb.c $(IINCDIR)/wrappers.h $(INCDIR)/dio.h $(INCDIR)/ihadcb.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -I$(IINCDIR) -c $<
 
 dioa.o: dioa.s
+	$(AS) $(ASFLAGS) -I$(MACDIR) $< >$*.lst
+
+mema.o: mema.s
 	$(AS) $(ASFLAGS) -I$(MACDIR) $< >$*.lst
