+++ src/dio.c
@@ -3,6 +3,7 @@
 #include <string.h>
 #include "asmdiocommon.h"
 #include "dio.h"
+#include "mem.h"
 #include "stow.h"
 #include "util.h"
 #include "s99.h"
@@ -120,59 +121,4 @@ unsigned DESERV(struct desp* PTR32 desp)
 int CLOSE(struct closecb* PTR32 closecb)
 {
   return CLOSEA(closecb);
-}
-void* PTR32 MALLOC24(size_t bytes)
-{
-  int ptr24;
-  ptr24 = MALOC24A(bytes);
-  void* PTR32 ptr = (void* PTR32) ptr24;
-  if (ptr24 == 0) {
-    fprintf(stderr, "Internal Error: Unable to allocate %d bytes below the bar\n", bytes);
-  } else {
-#ifdef DEBUG
-  memset(ptr, 0xFE, bytes);
-#else
-  memset(ptr, 0x00, bytes);
-#endif
-  }
-  return ptr;
-}
-
-int FREE24(void* PTR32 addr, size_t len)
-{
-  return FREE24A(addr, len);
-}
-
-void* PTR32 MALLOC31(size_t bytes)
-{
-  void* PTR32 p = __malloc31(bytes);
-  if (p == 0) {
-    fprintf(stderr, "Internal Error: Unable to allocate %d bytes below the line\n", bytes);
-  } else {
-#ifdef DEBUG
-  memset(p, 0xFE, bytes);
-#else
-  memset(p, 0x00, bytes);
-#endif
-  }
-  return p;
-}
-void FREE31(void* PTR32 addr)
-{
-  return free(addr);
-}
-
-void dumpstg(FILE* stream, void* p, size_t len)
-{
-  char* buff = p;
-  size_t i;
-  for (i=0; i<len; ++i) {
-    if ((i != 0) && (i % 16 == 0)) {
-      fprintf(stream, "\n");
-    }
-    if (i % 4 == 0) {
-      fprintf(stream, " ");
-    }
-    fprintf(stream, "%2.2X", buff[i]);
-  }
 }
