+++ src/s99.c
@@ -4,6 +4,7 @@
 #include <string.h>
 #include "asmdiocommon.h"
 #include "dio.h"
+#include "mem.h"
 #include "s99.h"
 #include "wrappers.h"
