+++ test/basicrddir.c
@@ -5,6 +5,7 @@
 
 #include "asmdiocommon.h"
 #include "dio.h"
+#include "mem.h"
 #include "ihadcb.h"
 #include "iosvcs.h"
 #include "s99.h"
