+++ test/basicread.c
@@ -4,6 +4,7 @@
 
 #include "asmdiocommon.h"
 #include "dio.h"
+#include "mem.h"
 #include "ihadcb.h"
 #include "iosvcs.h"
 #include "s99.h"
